# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/challenges": {
            "post": {
                "description": "Players can participate in an endless challenge by paying 20.01.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Challenges"
                ],
                "summary": "Participate in a Challenge",
                "parameters": [
                    {
                        "description": "Challenge Participation",
                        "name": "challenge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge started",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChallengeResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChallengeResponse"
                        }
                    }
                }
            }
        },
        "/challenges/results": {
            "get": {
                "description": "Retrieve a list of recent challenge results.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Challenges"
                ],
                "summary": "Get Recent Challenge Results",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of results to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of recent challenges",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Challenge"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/levels": {
            "get": {
                "description": "Retrieve a list of all levels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "Get all levels",
                "responses": {
                    "200": {
                        "description": "A list of levels",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Level"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new level in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "Add a new level",
                "parameters": [
                    {
                        "description": "Level Information",
                        "name": "level",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Level"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created level ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "description": "Retrieve a list of game logs with optional filters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Retrieve Game Logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by Player ID",
                        "name": "player_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Action Type (e.g., Register, Login)",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter logs from this time (RFC3339 format)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter logs up to this time (RFC3339 format)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of logs to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of game logs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Log"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new game operation log.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Create a Game Log",
                "parameters": [
                    {
                        "description": "Game Log Information",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New Log ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.LogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ChallengeRequest": {
            "type": "object",
            "required": [
                "player_id"
            ],
            "properties": {
                "player_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.ChallengeResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.LogRequest": {
            "type": "object",
            "required": [
                "action",
                "details",
                "player_id"
            ],
            "properties": {
                "action": {
                    "description": "e.g., Register, Login, Logout, Enter Room, Exit Room, Participate in Challenge, Challenge Result",
                    "type": "string"
                },
                "details": {
                    "description": "Additional information about the action",
                    "type": "string"
                },
                "player_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.LogResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Challenge": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "integer"
                },
                "won": {
                    "type": "boolean"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some thing wrong..."
                }
            }
        },
        "models.Level": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Log": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "e.g., Register, Login, Logout, Enter Room, Exit Room, Participate in Challenge, Challenge Result",
                    "type": "string"
                },
                "details": {
                    "description": "Additional information about the action",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Automatically set to current time on creation",
                    "type": "string"
                }
            }
        },
...
    }
}